+++ src/main/java/chess/Application.java
@@ -5,7 +5,6 @@ import chess.controller.ChessController;
 public class Application {
 
     public static void main(String[] args) {
-//        ChessController.run();
         ChessController.run();
     }
 }
