+++ src/main/java/chess/domain/piece/King.java
@@ -2,8 +2,8 @@ package chess.domain.piece;
 
 import chess.domain.Color;
 import chess.domain.Direction;
-import chess.domain.board.Square;
 import chess.domain.Type;
+import chess.domain.board.Square;
 
 import java.util.HashMap;
 import java.util.LinkedHashSet;
@@ -26,9 +26,7 @@ public class King extends Piece {
     }
 
     public static King of(Color color) {
-        if (color == null) {
-            throw new IllegalArgumentException("잘못된 입력입니다.");
-        }
+        checkColorNull(color);
         return CACHE.get(color.getName());
     }
