+++ src/test/java/chess/domain/board/SquareTest.java
@@ -1,6 +1,5 @@
-package chess.domain;
+package chess.domain.board;
 
-import chess.domain.board.Square;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
